move esg components to sub-folder
