+++ src/components/esg/Partners.vue
